+++ src/commands/hardis/project/audit/remotesites.ts
@@ -4,7 +4,9 @@ import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import * as fs from 'fs-extra';
 import * as glob from 'glob-promise';
+import * as psl from 'psl';
 import * as sortArray from 'sort-array';
+import * as url from 'url';
 import {
   catchMatches, generateReports
 } from '../../../../common/utils';
@@ -16,12 +18,12 @@ Messages.importMessagesDirectory(__dirname);
 // or any library that is using the messages framework can also be loaded this way.
 const messages = Messages.loadMessages('sfdx-hardis', 'org');
 
-export default class CallInCallOut extends SfdxCommand {
-  public static title = 'Audit CallIns and CallOuts';
+export default class RemoteSites extends SfdxCommand {
+  public static title = 'Audit Remote Sites';
 
-  public static description = messages.getMessage('auditCallInCallOut');
+  public static description = messages.getMessage('auditRemoteSites');
 
-  public static examples = ['$ sfdx hardis:project:audit:callouts'];
+  public static examples = ['$ sfdx hardis:project:audit:remotesites'];
 
   // public static args = [{name: 'file'}];
 
@@ -50,41 +52,25 @@ export default class CallInCallOut extends SfdxCommand {
   public async run(): Promise<AnyJson> {
     this.debug = this.flags.debug || false;
 
-    const pattern = '**/*.{cls,trigger}';
+    const pattern = '**/*.{remoteSite-meta.xml,remoteSite}';
     const catchers = [
       {
-        type: 'INBOUND',
-        subType: 'SOAP',
-        regex: /webservice static/gim,
+        type: '',
+        subType: '',
+        regex: /<url>(.*?)<\/url>/gim,
         detail: [
-          { name: 'webServiceName', regex: /webservice static (.*?){/gims }
-        ]
-      },
-      {
-        type: 'INBOUND',
-        subType: 'REST',
-        regex: /@RestResource/gim,
-        detail: [{ name: 'restResource', regex: /@RestResource\((.*?)\)/gims }]
-      },
-      {
-        type: 'OUTBOUND',
-        subType: 'HTTP',
-        regex: /new HttpRequest/gim,
-        detail: [
-          { name: 'endPoint', regex: /setEndpoint\((.*?);/gims },
-          { name: 'action', regex: /<soapenv:Body><[A-Za-z0-9_-]*:(.*?)>/gims}
+          { name: 'url', regex: /<url>(.*?)<\/url>/gims },
+          { name: 'active', regex: /<isActive>(.*?)<\/isActive>/gims },
+          { name: 'description', regex: /<description>(.*?)<\/description>/gimsu }
         ]
       }
     ];
-    const apexFiles = await glob(pattern);
+    const remoteSiteSettingsFiles = await glob(pattern);
     this.matchResults = [];
-    this.ux.log(`Browsing ${apexFiles.length} files`);
+    this.ux.log(`Browsing ${remoteSiteSettingsFiles.length} files`);
     // Loop in files
-    for (const file of apexFiles) {
+    for (const file of remoteSiteSettingsFiles) {
       const fileText = await fs.readFile(file, 'utf8');
-      if (fileText.startsWith('hidden') || fileText.includes('@isTest')) {
-        continue;
-      }
       // Loop on criteria to find matches in this file
       for (const catcher of catchers) {
         const catcherMatchResults = await catchMatches(catcher, file, fileText, this);
@@ -95,55 +81,45 @@ export default class CallInCallOut extends SfdxCommand {
     // Format result
     const result: any[] = this.matchResults.map((item: any) => {
       return {
-        type: item.type,
-        subType: item.subType,
+        name: item.fileName.replace('.remoteSite-meta.xml', '').replace('.remoteSite', ''),
         fileName: item.fileName,
         nameSpace: item.fileName.includes('__')
           ? item.fileName.split('__')[0]
           : 'Custom',
         matches: item.matches,
-        detail:
-          Object.keys(item.detail)
-            .map(
-              (key: string) =>
-                key +
-                ': ' +
-                item.detail[key]
-                  .map(
-                    (extractedText: string) =>
-                      extractedText
-                        .replace(/(\r\n|\n|\r)/gm, '') // Remove new lines from result
-                        .replace(/\s+/g, ' ') // Replace multiple whitespaces by single whitespaces
-                  )
-                  .join(' | ')
-            )
-            .join(' || ') || ''
+        url: (item.detail?.url) ? item.detail.url[0] : '',
+        active: (item.detail?.active) ? 'yes' : 'no',
+        description: (item.detail?.description) ? item.detail.description[0] : '',
+        protocol: (item.detail.url[0].includes('https')) ? 'HTTPS' : 'HTTP',
+        domain: psl.parse(new url.URL(item.detail.url[0]).hostname).domain
       };
     });
 
     // Sort array
     const resultSorted = sortArray(result, {
-      by: ['type', 'subType', 'fileName', 'matches'],
-      order: ['asc', 'asc', 'asc', 'desc']
+      by: ['protocol', 'domain', 'name', 'active', 'description'],
+      order: ['asc', 'asc', 'asc', 'desc', 'asc']
     });
 
     // Display as table
     const resultsLight = JSON.parse(JSON.stringify(resultSorted));
     console.table(
       resultsLight.map((item: any) => {
+        delete item.fileName;
         delete item.detail;
+        delete item.matches;
         return item;
       })
     );
 
-    // Generate output files
+    // Export into csv & excel file
     const columns = [
-      { key: 'type', header: 'IN/OUT' },
-      { key: 'subType', header: 'Protocol' },
-      { key: 'fileName', header: 'Apex' },
-      { key: 'nameSpace', header: 'Namespace' },
-      { key: 'matches', header: 'Number'},
-      { key: 'detail', header: 'Detail' }
+      { key: 'protocol', header: 'Protocol' },
+      { key: 'domain', header: 'Domain' },
+      { key: 'name', header: 'Name' },
+      { key: 'url', header: 'URL' },
+      { key: 'active', header: 'Active' },
+      { key: 'description', header: 'Description' }
     ];
     const reportFiles = await generateReports(resultSorted, columns, this);